fix csrf typo, remove plugins and refactor when we have more things.
