+++ server/src/controllers/auth.controller.ts
@@ -1,4 +1,5 @@
 import { HttpContext } from '@kondah/http-context'
+import csurf from 'csurf'
 import { Controller, Delete, Get, Middleware } from '@kondah/http-controller'
 
 import { Request, Response } from 'express'
