--- server/src/plugins/README.md
@@ -1,3 +1,0 @@
-# Description
-
-This is the first time using Kondah(https://kondah.dev/) which is a micro framework built by me. In this project I've chosen to really over-do the idea of **plugins** to get a good idea on what might be a potential problem for future projects.
